Add scripts
